app/models/__init__.py
# models/__init__.py
from sqlalchemy import MetaData

metadata = MetaData()


from ._base import BaseModel

from .delivery_station import DeliveryStation
from .department import Department
from .department_order import DepartmentOrder
from .department_order_item import DepartmentOrderItem
from .item import Item
from .kiosk import Kiosk
from .order import Order
from .printer import Printer
from .user import User
