requirements.txt
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy-Utils==0.41.2
passlib==1.7.4
Pillow==11.1.0
# dotenv
# psycopg2==2.9.5
# setuptools==65.5.1
# wheel==0.38.4
# Werkzeug==2.2.2
# sqlalchemy-utils==0.41.2
